+++ texthub/modules/det_heads/__init__.py
@@ -1,3 +1,4 @@
 from .pan_head import PanHead
 from .db_head import DBHead
 from .pse_head import PseHead
+from .pa_head import PanCPPHead